Fix more warnings and make a shared scheme for Carthage support.

+++ SocketRocket/SocketRocket/base64.c
@@ -129,8 +129,7 @@ static const char Pad64 = '=';
  */
 
 #if !defined(HAVE_B64_NTOP) && !defined(HAVE___B64_NTOP) 
-int
-b64_ntop(u_char const *src, size_t srclength, char *target, size_t targsize)
+int b64_ntop(u_char const *src, size_t srclength, char *target, size_t targsize)
 {
 	size_t datalength = 0;
 	u_char input[3];
@@ -180,7 +179,7 @@ b64_ntop(u_char const *src, size_t srclength, char *target, size_t targsize)
 	if (datalength >= targsize)
 		return (-1);
 	target[datalength] = '\0';	/* Returned value doesn't count \0. */
-	return (datalength);
+	return (int)datalength;
 }
 #endif /* !defined(HAVE_B64_NTOP) && !defined(HAVE___B64_NTOP) */
